requirements.txt
python-dotenv 
openai
aiohttp 
fastapi[all]